Se hizo la parte de los Roles y Permisos y se dejo implementado la parte grafica del sistema

+++ app/Http/Controllers/UsuarioController.php
@@ -6,19 +6,21 @@ use App\Http\Requests\UserCreateRequest;
 use App\Http\Requests\UserEditRequest;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Spatie\Permission\Models\Role;
 
 class UsuarioController extends Controller
 {
     public function index(){
 
-        $usuarios = User::paginate(5);
+        $usuarios = User::paginate(8);
         return view('usuarios.index', compact('usuarios'));
 
     }
 
     public function create(){
 
-        return view('usuarios.create');
+        $roles = Role::all()->pluck('name', 'id');
+        return view('usuarios.create', compact('roles'));
 
     }
 
@@ -31,25 +33,31 @@ class UsuarioController extends Controller
 
         // ]);
 
-
-        User::create($request->only('name','email')
+        $usuario = User::create($request->only('name','email')
         + [
             'password' => bcrypt($request->input('password')),
         ]);
-        return redirect()->route('usuarios.index')->with('success', 'usuario creado correctamente');
+
+        $roles = $request->input('roles', []);
+        $usuario->syncRoles($roles);
+        return redirect()->route('usuarios.index', $usuario->id)->with('success', 'usuario creado correctamente');
         // return redirect()->back(); // QUE CUANDO CREAA NOS REDIRECCIONE A LA VITA
 
     }
 
     public function show(User $usuario){
         //$usuario = User::findOrFail($id);
+        // load para cargar relaciones de uno a muchos o de muchos a muchos
+        $usuario->load('roles');
         return view('usuarios.show', compact('usuario'));
 
     }
 
     public function edit(User $usuario){
 
-        return view('usuarios.edit', compact('usuario'));
+        $roles = Role::all()->pluck('name', 'id');
+        $usuario->load('roles');
+        return view('usuarios.edit', compact('usuario', 'roles'));
 
     }
 
@@ -69,10 +77,19 @@ class UsuarioController extends Controller
         // }
 
         $usuario->update($data);
+
+        $roles = $request->input('roles', []);
+        $usuario->syncRoles($roles);
         return redirect()->route('usuarios.index')->with('success', 'Usuario actualizado correctamente');
     }
 
     public function destroy(User $usuario){
+
+
+
+        if (auth()->user()->id == $usuario->id) {
+            return redirect()->route('usuarios.index')->with('success', 'No se puede eliminar a usted mismo ');
+        }
 
         $usuario->delete();
         return back()->with('success', 'Usuario eliminado correctamente');
